Move init command steps to init_command.py
